-- did shadow mapping :)

+++ 19-Shadow_mapping/shaders/fragmentShader_shadow.glsl
@@ -1,7 +1,6 @@
-precision mediump float;
+precision highp float;
 
 varying float v_Depth;
-
 void main() {
-    gl_FragColor = vec4(v_Depth, 0.0, 0.0, 1.0);
+    gl_FragColor =  vec4(v_Depth,0.0,0.0,1.0);
 }
